more data source methods, insert and delete commands, zones and buttons

+++ app/src/main/java/com/sourceallies/android/zonebeacon/data/model/Command.java
@@ -46,7 +46,7 @@ public class Command implements DatabaseTable {
 
     @Setter
     @Getter
-    private int id;
+    private long id;
 
     @Setter
     @Getter
@@ -94,7 +94,7 @@ public class Command implements DatabaseTable {
             String column = cursor.getColumnName(i);
 
             if (column.equals(COLUMN_ID)) {
-                setId(cursor.getInt(i));
+                setId(cursor.getLong(i));
             } else if (column.equals(COLUMN_NAME)) {
                 setName(cursor.getString(i));
             } else if (column.equals(COLUMN_GATEWAY_ID)) {
